refactoring and adding some javadoc
