+++ src/main/java/com/primer/tokeniser/dto/SaleDTO.java
@@ -1,4 +1,4 @@
-package com.primer.tokeniser.web.rest;
+package com.primer.tokeniser.dto;
 
 import java.math.BigDecimal;
