--- src/main/java/com/primer/tokeniser/web/rest/errors/SaleFailedException.java
@@ -1,10 +1,0 @@
-package com.primer.tokeniser.web.rest.errors;
-
-import com.braintreegateway.exceptions.BraintreeException;
-
-public class SaleFailedException extends BraintreeException {
-
-    public SaleFailedException(final String message) {
-        super(message);
-    }
-}
